[Refactor] Extract SegmentPKIterator into its own file
Move SegmentPKIterator and its SegmentPKChunkRef return type out of
rowset_update_state.{h,cpp} into a dedicated segment_pk_iterator.{h,cpp}.
rowset_update_state.h re-includes the new header so existing consumers
are unaffected. Pure code motion; no behavior change.
